Refactored into tasks pattern
